requirements.txt
Flask==2.0.2
Flask-PyMongo==2.3.0
PyJWT==2.1.0
Werkzeug==2.0.2
python-dotenv